cloud docker

+++ cloud/docker/deploy.sh
@@ -1,5 +1,21 @@
 #!/bin/bash
 
+serverUrl=""
+if [ -e "SERVER_URL" ]; then
+    serverUrl=$(cat "SERVER_URL")
+fi
+
+if [ -n "$1" ]; then
+    serverUrl="$1"
+fi
+
+if [ -z "$serverUrl" ]; then
+    echo "Enter server URL!"
+    exit 1
+fi
+
+echo $serverUrl > ./SERVER_URL
+
 if [ -d "./budgi" ]; then
     # Delete old directory
     rm -rf ./budgi/
@@ -23,9 +39,11 @@ else
     docker compose down
 
     cp -rf budgi/VERSION ./VERSION
-    cp -rf budgi/cloud/gilbva/cfg/ ./cfg/
-    cp -rf budgi/cloud/gilbva/compose.yml ./compose.yml
+    cp -rf budgi/cloud/docker/cfg/ ./cfg/
+    cp -rf budgi/cloud/docker/compose.yml ./compose.yml
     rm -rf ./budgi/
+
+    sed -i "s/SERVER_URL/$serverUrl/g" ./cfg/nginx/budgi.conf
 
     docker compose pull
     docker compose up -d
